indexer/requirements.txt
eventlet==0.19.0
greenlet==0.4.10
peewee==2.8.3
pika==0.10.0
psycopg2==2.6.2
sparql-client==2.5
